Removing 'extra' and separating "links" in the definition
If we set the presentation rules in the trigger or provide details for the
delivery methods, then there's no need for "extras" anymore. Extra field
is used for the PresentationModel, it sends details from the trigger context
into the PresentationModel to create dynamic results, but unless I'm missing
something, the results can be provided on trigger -- with the option to set two
different options for the delivery method in case needed (like separate messages
or links depending on whether the notification is bundled).