Add a question around truly dynamic title
Bundled notifications often use a dynamic number in the header message that
checks how many notifications are in each bundle. This means we can't pre-produce
the message on trigger, but it also raises a question about whether the underlying
system can do this at all, given that bundles are only really respected/rendered in
specific deliveries (the web popup).